pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "power-fair-value"
version = "0.1.0"
description = "Power market fair value forecasting pipeline"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "pyarrow>=12.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "matplotlib>=3.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.8.0",
]

[project.scripts]
power-fair-value = "pipeline.cli:main"

[tool.setuptools]
packages = ["pipeline", "ingest", "qa", "features", "models", "validation", "trading", "reporting", "utils"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py39']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
